Extracted YML persistence functionality
